pyproject.toml
[project]
name = "katann-web"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "python-socketio>=5.16.0",
    "uvicorn>=0.40.0",
]
